Refactored some file ops
