+++ dependencies/engine/filesystem/include/win32/filesystem/File.hpp
