+++ dependencies/engine/resources/source/common/ModelLoader.cpp
@@ -3,7 +3,7 @@
 #include "resources/Model.hpp"
 #include "utilities/IOC.hpp"
 #include "filesystem/IFileSystem.hpp"
-#include "File.hpp"
+#include "filesystem/File.hpp"
 #include <vector>
 
 using namespace std;
